+++ xobliam/fetcher/__init__.py
@@ -3,6 +3,7 @@
 from .cache import MessageCache
 from .label_ops import (
     apply_label_to_messages,
+    create_filter_for_senders,
     create_label,
     get_label_id_by_name,
     merge_labels,
@@ -13,6 +14,7 @@ from .messages import fetch_messages
 __all__ = [
     "MessageCache",
     "apply_label_to_messages",
+    "create_filter_for_senders",
     "create_label",
     "fetch_labels",
     "fetch_messages",